培训计划清单接口

+++ dao/dao_BASF/src/main/java/com/nt/dao_BASF/Startprogram.java
@@ -74,7 +74,7 @@ public class Startprogram extends BaseModel {
     private String isonline;
 
     //培训时长
-    private Integer thelength;
+    private Double thelength;
 
     //培训有效期（月）
     private Integer validity;
